+++ common/home-manager-sets/hyprland-desktop.nix
@@ -1,4 +1,4 @@
 { self }:
 let inherit (self.common.home-manager-module-sets) hyprland-desktop-minimal;
 in hyprland-desktop-minimal
-++ (with self.homeManagerModules; [ desktop-packages-extra ])
+++ (with self.homeManagerModules; [ desktop-packages-extra mako ])