+++ common/home-manager-sets/hyprland-ironbar-desktop.nix
@@ -1,0 +1,3 @@
+{ self }:
+let inherit (self.common.home-manager-module-sets) hyprland-desktop;
+in hyprland-desktop ++ (with self.homeManagerModules; [ ironbar ])