+++ hosts/alakazam/default.nix
@@ -1,7 +1,8 @@
 { config, pkgs, self, ... }:
 
 let
-  inherit (self.common.home-manager-module-sets) base hyprland-desktop games;
+  inherit (self.common.home-manager-module-sets)
+    base hyprland-ironbar-desktop games;
   inherit (self.lib) merge;
 
   inherit (pkgs) system;
@@ -14,8 +15,7 @@ let
 
   jay = self.common.users.jay {
     inherit config pkgs;
-    modules = hyprland-desktop ++ games
-      ++ (with self.homeManagerModules; [ ironbar ]);
+    modules = hyprland-ironbar-desktop ++ games;
   };
 
   user-configs = merge [ builder jay ];